Revert "Refactor internal/os into x/wasm"
This reverts commit 200df684987e354ae2c27f59330c4bfd7ef40fa6.
